refactor: move `formatWithValidation` next to `formatUrl`
